+++ src/main/java/com/onseo/TestTask/domain/dto/Album.java
@@ -1,4 +1,4 @@
-package com.exadel.development.TestOnseo.domain.dto;
+package com.onseo.TestTask.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +7,8 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class ToDo {
+public class Album {
     private Long userId;
     private Long id;
     private String title;
-    private boolean completed;
 }
