+++ src/main/java/com/onseo/TestTask/domain/dto/Post.java
@@ -1,4 +1,4 @@
-package com.exadel.development.TestOnseo.domain.dto;
+package com.onseo.TestTask.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,10 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class Photo {
-    private Long albumId;
+public class Post {
+    private Long userId;
     private Long id;
     private String title;
-    private String url;
-    private String thumbnailUrl;
+    private String body;
 }
