+++ src/main/java/com/onseo/TestTask/domain/dto/ToDo.java
@@ -1,4 +1,4 @@
-package com.exadel.development.TestOnseo.domain.dto;
+package com.onseo.TestTask.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,8 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class Album {
+public class ToDo {
     private Long userId;
     private Long id;
     private String title;
+    private boolean completed;
 }
