+++ src/main/java/com/onseo/TestTask/service/exceptions/DataLoadingException.java
@@ -1,4 +1,4 @@
-package com.exadel.development.TestOnseo.service.exceptions;
+package com.onseo.TestTask.service.exceptions;
 
 public class DataLoadingException extends RuntimeException {
     public DataLoadingException(String message, Throwable cause) {
