+++ src/main/java/com/onseo/TestTask/service/exceptions/DataMappingException.java
@@ -1,4 +1,4 @@
-package com.exadel.development.TestOnseo.service.exceptions;
+package com.onseo.TestTask.service.exceptions;
 
 public class DataMappingException extends RuntimeException {
     public DataMappingException(String message, Throwable cause) {
